+++ xmtp_api_d14n/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "xmtp_api_d14n"
-edition = "2021"
+edition = "2024"
 license.workspace = true
 version.workspace = true
